Make organization of platform mixin packages consistent
Fixes #2688
