--- fabric/src/main/java/net/caffeinemc/mods/sodium/mixin/fabric/features/model/MultiPartBakedModelMixin.java
@@ -1,78 +1,0 @@
-package net.caffeinemc.mods.sodium.mixin.fabric.features.model;
-
-import it.unimi.dsi.fastutil.objects.Reference2ReferenceOpenHashMap;
-import net.minecraft.client.renderer.block.model.BakedQuad;
-import net.minecraft.client.resources.model.BakedModel;
-import net.minecraft.client.resources.model.MultiPartBakedModel;
-import net.minecraft.core.Direction;
-import net.minecraft.util.RandomSource;
-import net.minecraft.world.level.block.state.BlockState;
-import org.apache.commons.lang3.tuple.Pair;
-import org.spongepowered.asm.mixin.*;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.locks.StampedLock;
-import java.util.function.Predicate;
-
-@Mixin(MultiPartBakedModel.class)
-public class MultiPartBakedModelMixin {
-    @Unique
-    private final Map<BlockState, BakedModel[]> stateCacheFast = new Reference2ReferenceOpenHashMap<>();
-    @Unique
-    private final StampedLock lock = new StampedLock();
-
-    @Shadow
-    @Final
-    private List<Pair<Predicate<BlockState>, BakedModel>> selectors;
-
-    /**
-     * @author JellySquid
-     * @reason Avoid expensive allocations and replace bitfield indirection
-     */
-    @Overwrite
-    public List<BakedQuad> getQuads(BlockState state, Direction face, RandomSource random) {
-        if (state == null) {
-            return Collections.emptyList();
-        }
-
-        BakedModel[] models;
-
-        long readStamp = this.lock.readLock();
-        try {
-            models = this.stateCacheFast.get(state);
-        } finally {
-            this.lock.unlockRead(readStamp);
-        }
-
-        if (models == null) {
-            long writeStamp = this.lock.writeLock();
-            try {
-                List<BakedModel> modelList = new ArrayList<>(this.selectors.size());
-
-                for (Pair<Predicate<BlockState>, BakedModel> pair : this.selectors) {
-                    if (pair.getLeft().test(state)) {
-                        modelList.add(pair.getRight());
-                    }
-                }
-
-                models = modelList.toArray(BakedModel[]::new);
-                this.stateCacheFast.put(state, models);
-            } finally {
-                this.lock.unlockWrite(writeStamp);
-            }
-        }
-
-        List<BakedQuad> quads = new ArrayList<>();
-        long seed = random.nextLong();
-
-        for (BakedModel model : models) {
-            random.setSeed(seed);
-            quads.addAll(model.getQuads(state, face, random));
-        }
-
-        return quads;
-    }
-}
