+++ fabric/src/main/resources/sodium-fabric.mixins.json
@@ -1,5 +1,5 @@
 {
-  "package": "net.caffeinemc.mods.sodium.mixin.fabric",
+  "package": "net.caffeinemc.mods.sodium.mixin",
   "required": true,
   "compatibilityLevel": "JAVA_17",
   "injectors": {
@@ -9,10 +9,10 @@
     "conformVisibility": true
   },
   "client": [
-    "core.model.quad.BakedQuadMixin",
-    "features.model.MultiPartBakedModelMixin",
-    "features.model.WeightedBakedModelMixin",
-    "features.render.model.block.ModelBlockRendererMixin",
-    "features.world.biome.BiomeMixin"
+      "core.model.quad.BakedQuadMixin",
+      "features.model.MultiPartBakedModelMixin",
+      "features.model.WeightedBakedModelMixin",
+      "features.render.model.block.ModelBlockRendererMixin",
+      "features.world.biome.BiomeMixin"
   ]
 }
