+++ neoforge/build.gradle.kts
@@ -139,3 +139,4 @@ tasks {
         from(configurationCommonModResources)
     }
 }
+
