+++ neoforge/src/main/java/net/caffeinemc/mods/sodium/mixin/features/render/model/block/ModelBlockRendererMixin.java
@@ -1,0 +1,93 @@
+package net.caffeinemc.mods.sodium.mixin.features.render.model.block;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import net.caffeinemc.mods.sodium.api.util.ColorABGR;
+import net.caffeinemc.mods.sodium.api.vertex.buffer.VertexBufferWriter;
+import net.caffeinemc.mods.sodium.client.model.quad.BakedQuadView;
+import net.caffeinemc.mods.sodium.client.render.immediate.model.BakedModelEncoder;
+import net.caffeinemc.mods.sodium.client.render.texture.SpriteUtil;
+import net.caffeinemc.mods.sodium.client.render.vertex.VertexConsumerUtils;
+import net.caffeinemc.mods.sodium.client.util.DirectionUtil;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.block.ModelBlockRenderer;
+import net.minecraft.client.renderer.block.model.BakedQuad;
+import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.core.Direction;
+import net.minecraft.util.Mth;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.levelgen.SingleThreadedRandomSource;
+import net.neoforged.neoforge.client.model.data.ModelData;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.List;
+
+@Mixin(ModelBlockRenderer.class)
+public class ModelBlockRendererMixin {
+    @Unique
+    private final RandomSource random = new SingleThreadedRandomSource(42L);
+
+    @Unique
+    @SuppressWarnings("ForLoopReplaceableByForEach")
+    private static void renderQuads(PoseStack.Pose matrices, VertexBufferWriter writer, int defaultColor, List<BakedQuad> quads, int light, int overlay) {
+        for (int i = 0; i < quads.size(); i++) {
+            BakedQuad bakedQuad = quads.get(i);
+
+            if (bakedQuad.getVertices().length < 32) {
+                continue; // ignore bad quads
+            }
+
+            BakedQuadView quad = (BakedQuadView) bakedQuad;
+
+            int color = quad.hasColor() ? defaultColor : 0xFFFFFFFF;
+
+            BakedModelEncoder.writeQuadVertices(writer, matrices, quad, color, light, overlay, false);
+
+            SpriteUtil.markSpriteActive(quad.getSprite());
+        }
+    }
+
+    /**
+     * @reason Use optimized vertex writer intrinsics, avoid allocations
+     * @author JellySquid
+     */
+    @Inject(method = "renderModel(Lcom/mojang/blaze3d/vertex/PoseStack$Pose;Lcom/mojang/blaze3d/vertex/VertexConsumer;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/client/resources/model/BakedModel;FFFIILnet/neoforged/neoforge/client/model/data/ModelData;Lnet/minecraft/client/renderer/RenderType;)V", at = @At("HEAD"), cancellable = true)
+    private void renderFast(PoseStack.Pose entry, VertexConsumer vertexConsumer, BlockState blockState, BakedModel bakedModel, float red, float green, float blue, int light, int overlay, ModelData modelData, RenderType renderType, CallbackInfo ci) {
+        var writer = VertexConsumerUtils.convertOrLog(vertexConsumer);
+        if (writer == null) {
+            return;
+        }
+
+        ci.cancel();
+
+        RandomSource random = this.random;
+
+        // Clamp color ranges
+        red = Mth.clamp(red, 0.0F, 1.0F);
+        green = Mth.clamp(green, 0.0F, 1.0F);
+        blue = Mth.clamp(blue, 0.0F, 1.0F);
+
+        int defaultColor = ColorABGR.pack(red, green, blue, 1.0F);
+
+        for (Direction direction : DirectionUtil.ALL_DIRECTIONS) {
+            random.setSeed(42L);
+            List<BakedQuad> quads = bakedModel.getQuads(blockState, direction, random, modelData, renderType);
+
+            if (!quads.isEmpty()) {
+                renderQuads(entry, writer, defaultColor, quads, light, overlay);
+            }
+        }
+
+        random.setSeed(42L);
+        List<BakedQuad> quads = bakedModel.getQuads(blockState, null, random, modelData, renderType);
+
+        if (!quads.isEmpty()) {
+            renderQuads(entry, writer, defaultColor, quads, light, overlay);
+        }
+    }
+}
