--- neoforge/src/main/java/net/caffeinemc/mods/sodium/neoforge/mixin/AuxiliaryLightManagerMixin.java
@@ -1,9 +1,0 @@
-package net.caffeinemc.mods.sodium.neoforge.mixin;
-
-import net.caffeinemc.mods.sodium.client.world.SodiumAuxiliaryLightManager;
-import net.neoforged.neoforge.common.world.AuxiliaryLightManager;
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(AuxiliaryLightManager.class)
-public interface AuxiliaryLightManagerMixin extends SodiumAuxiliaryLightManager {
-}
