--- neoforge/src/main/java/net/caffeinemc/mods/sodium/neoforge/mixin/ChunkRenderTypeSetAccessor.java
@@ -1,19 +1,0 @@
-package net.caffeinemc.mods.sodium.neoforge.mixin;
-
-import net.neoforged.neoforge.client.ChunkRenderTypeSet;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-import org.spongepowered.asm.mixin.gen.Invoker;
-
-import java.util.BitSet;
-
-@Mixin(ChunkRenderTypeSet.class)
-public interface ChunkRenderTypeSetAccessor {
-    @Invoker("<init>")
-    static ChunkRenderTypeSet create(BitSet set) {
-        throw new IllegalStateException("Not shadowed");
-    }
-
-    @Accessor
-    BitSet getBits();
-}
