+++ neoforge/src/main/resources/sodium-neoforge.mixins.json
@@ -1,5 +1,5 @@
 {
-  "package" : "net.caffeinemc.mods.sodium.neoforge.mixin",
+  "package" : "net.caffeinemc.mods.sodium.mixin",
   "required" : true,
   "compatibilityLevel" : "JAVA_17",
   "injectors" : {
@@ -9,19 +9,19 @@
     "conformVisibility" : true
   },
   "client" : [
-    "core.model.quad.BakedQuadMixin",
-    "features.model.MultiPartBakedModelMixin",
-    "features.model.WeightedBakedModelMixin",
-    "features.render.model.block.ModelBlockRendererMixin",
-    "features.world.biome.BiomeMixin",
-    "AbstractBlockRenderContextMixin",
-    "AuxiliaryLightManagerMixin",
-    "ChunkRenderTypeSetAccessor",
-    "ClientHooksMixin",
-    "EntrypointMixin",
-    "LevelSliceMixin",
-    "ModelDataMixin",
-    "ResourcePackLoaderMixin",
-    "SimpleBakedModelAccessor"
+      "core.model.quad.BakedQuadMixin",
+      "features.model.MultiPartBakedModelMixin",
+      "features.model.WeightedBakedModelMixin",
+      "features.render.model.block.ModelBlockRendererMixin",
+      "features.world.biome.BiomeMixin",
+      "platform.neoforge.AbstractBlockRenderContextMixin",
+      "platform.neoforge.AuxiliaryLightManagerMixin",
+      "platform.neoforge.ChunkRenderTypeSetAccessor",
+      "platform.neoforge.ClientHooksMixin",
+      "platform.neoforge.EntrypointMixin",
+      "platform.neoforge.LevelSliceMixin",
+      "platform.neoforge.ModelDataMixin",
+      "platform.neoforge.ResourcePackLoaderMixin",
+      "platform.neoforge.SimpleBakedModelAccessor"
   ]
 }
